+++ lib/gridstore/src/gridstore/mod.rs
@@ -1,6 +1,6 @@
+mod append_only;
 mod dynamic;
 mod reader;
-mod serverless;
 pub(crate) mod view;
 
 #[cfg(test)]
@@ -8,6 +8,7 @@ mod tests;
 
 use std::path::PathBuf;
 
+use append_only::AppendOnlyGridstore;
 use common::counter::counter_cell::CounterCell;
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::counter::referenced_counter::HwMetricRefCounter;
@@ -16,7 +17,6 @@ use common::universal_io::{MmapFile, Populate, UniversalWrite, UniversalWriteFil
 use dynamic::DynamicGridstore;
 use reader::CONFIG_FILENAME;
 pub use reader::GridstoreReader;
-use serverless::ServerlessGridstore;
 pub use view::GridstoreView;
 
 use crate::Result;
@@ -34,7 +34,7 @@ pub type Flusher = Box<dyn FnOnce() -> std::result::Result<(), GridstoreError> +
 /// Operates in one of two modes, specified on creation and automatically selected when opening:
 ///
 /// - [`Mode::Dynamic`]: values can be updated and deleted, freed blocks are tracked and reused.
-/// - [`Mode::Serverless`]: append-only variant for serverless deployments, values cannot be
+/// - [`Mode::AppendOnly`]: append-only variant for serverless deployments, values cannot be
 ///   updated or deleted, and must be put in monotonically increasing point offset order.
 ///
 /// Assumes sequential IDs to the values (0, 1, 2, 3, ...)
@@ -53,7 +53,7 @@ where
     S: UniversalWrite + 'static,
 {
     Dynamic(DynamicGridstore<V, S>),
-    Serverless(ServerlessGridstore<V, S>),
+    AppendOnly(AppendOnlyGridstore<V, S>),
 }
 
 impl<V, S> Gridstore<V, S>
@@ -65,14 +65,14 @@ where
     pub fn files(&self) -> Vec<PathBuf> {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.files(),
-            GridstoreVariant::Serverless(storage) => storage.files(),
+            GridstoreVariant::AppendOnly(storage) => storage.files(),
         }
     }
 
     pub fn immutable_files(&self) -> Vec<PathBuf> {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.immutable_files(),
-            GridstoreVariant::Serverless(storage) => storage.immutable_files(),
+            GridstoreVariant::AppendOnly(storage) => storage.immutable_files(),
         }
     }
 
@@ -107,12 +107,12 @@ where
                     variant: GridstoreVariant::Dynamic(storage),
                 })
             }
-            // The serverless mode does not use the universal io backend, it reads and writes
+            // The append-only mode does not use the universal io backend, it reads and writes
             // files directly
-            Mode::Serverless => {
-                let storage = ServerlessGridstore::new(base_path, options)?;
+            Mode::AppendOnly => {
+                let storage = AppendOnlyGridstore::new(base_path, options)?;
                 Ok(Self {
-                    variant: GridstoreVariant::Serverless(storage),
+                    variant: GridstoreVariant::AppendOnly(storage),
                 })
             }
         }
@@ -130,12 +130,12 @@ where
                     variant: GridstoreVariant::Dynamic(storage),
                 })
             }
-            // The serverless mode does not use the universal io backend, it reads and writes
+            // The append-only mode does not use the universal io backend, it reads and writes
             // files directly
-            Mode::Serverless => {
-                let storage = ServerlessGridstore::open(base_path, config)?;
+            Mode::AppendOnly => {
+                let storage = AppendOnlyGridstore::open(base_path, config)?;
                 Ok(Self {
-                    variant: GridstoreVariant::Serverless(storage),
+                    variant: GridstoreVariant::AppendOnly(storage),
                 })
             }
         }
@@ -145,7 +145,7 @@ where
     ///
     /// Returns true if the value existed previously and was updated, false if it was newly inserted.
     ///
-    /// In serverless mode values must be put at monotonically increasing point offsets, and
+    /// In append-only mode values must be put at monotonically increasing point offsets, and
     /// cannot be overwritten.
     pub fn put_value(
         &mut self,
@@ -157,7 +157,7 @@ where
             GridstoreVariant::Dynamic(storage) => {
                 storage.put_value(point_offset, value, hw_counter)
             }
-            GridstoreVariant::Serverless(storage) => {
+            GridstoreVariant::AppendOnly(storage) => {
                 storage.put_value(point_offset, value, hw_counter)
             }
         }
@@ -168,11 +168,11 @@ where
     /// Returns None if the point_offset, page, or value was not found.
     /// Returns the deleted value otherwise.
     ///
-    /// Not supported in serverless mode, returns an error.
+    /// Not supported in append-only mode, returns an error.
     pub fn delete_value(&mut self, point_offset: PointOffset) -> Result<Option<V>> {
         match &mut self.variant {
             GridstoreVariant::Dynamic(storage) => storage.delete_value(point_offset),
-            GridstoreVariant::Serverless(storage) => storage.delete_value(point_offset),
+            GridstoreVariant::AppendOnly(storage) => storage.delete_value(point_offset),
         }
     }
 
@@ -182,7 +182,7 @@ where
     pub fn clear(&mut self) -> Result<()> {
         match &mut self.variant {
             GridstoreVariant::Dynamic(storage) => storage.clear(),
-            GridstoreVariant::Serverless(storage) => storage.clear(),
+            GridstoreVariant::AppendOnly(storage) => storage.clear(),
         }
     }
 
@@ -193,7 +193,7 @@ where
     pub fn wipe(self) -> Result<()> {
         match self.variant {
             GridstoreVariant::Dynamic(storage) => storage.wipe(),
-            GridstoreVariant::Serverless(storage) => storage.wipe(),
+            GridstoreVariant::AppendOnly(storage) => storage.wipe(),
         }
     }
 
@@ -201,7 +201,7 @@ where
     pub fn get_storage_size_bytes(&self) -> Result<usize> {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.get_storage_size_bytes(),
-            GridstoreVariant::Serverless(storage) => storage.get_storage_size_bytes(),
+            GridstoreVariant::AppendOnly(storage) => storage.get_storage_size_bytes(),
         }
     }
 
@@ -212,7 +212,7 @@ where
     ) -> Result<Option<V>> {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.get_value::<P>(point_offset, hw_counter),
-            GridstoreVariant::Serverless(storage) => {
+            GridstoreVariant::AppendOnly(storage) => {
                 storage.get_value::<P>(point_offset, hw_counter)
             }
         }
@@ -236,7 +236,7 @@ where
             GridstoreVariant::Dynamic(storage) => {
                 storage.read_values::<P, U, E>(point_offsets, callback, hw_counter_cell)
             }
-            GridstoreVariant::Serverless(storage) => {
+            GridstoreVariant::AppendOnly(storage) => {
                 storage.read_values::<P, U, E>(point_offsets, callback, hw_counter_cell)
             }
         }
@@ -246,14 +246,14 @@ where
     pub fn get_pointer(&self, point_offset: PointOffset) -> Option<ValuePointer> {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.get_pointer(point_offset),
-            GridstoreVariant::Serverless(storage) => storage.get_pointer(point_offset),
+            GridstoreVariant::AppendOnly(storage) => storage.get_pointer(point_offset),
         }
     }
 
     pub fn max_point_offset(&self) -> PointOffset {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.max_point_offset(),
-            GridstoreVariant::Serverless(storage) => storage.max_point_offset(),
+            GridstoreVariant::AppendOnly(storage) => storage.max_point_offset(),
         }
     }
 
@@ -267,7 +267,7 @@ where
     {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.iter(callback, hw_counter),
-            GridstoreVariant::Serverless(storage) => storage.iter(callback, hw_counter),
+            GridstoreVariant::AppendOnly(storage) => storage.iter(callback, hw_counter),
         }
     }
 }
@@ -277,27 +277,27 @@ impl<V, S: UniversalWrite + 'static> Gridstore<V, S> {
     pub fn flusher(&self) -> Flusher {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.flusher(),
-            GridstoreVariant::Serverless(storage) => storage.flusher(),
+            GridstoreVariant::AppendOnly(storage) => storage.flusher(),
         }
     }
 
     /// Populate all parts of the storage in the mmap.
     ///
-    /// No-op in serverless mode, which does not memory map its files.
+    /// No-op in append-only mode, which does not memory map its files.
     pub fn populate(&self) -> Result<()> {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.populate(),
-            GridstoreVariant::Serverless(storage) => storage.populate(),
+            GridstoreVariant::AppendOnly(storage) => storage.populate(),
         }
     }
 
     /// Drop disk cache.
     ///
-    /// No-op in serverless mode, which does not memory map its files.
+    /// No-op in append-only mode, which does not memory map its files.
     pub fn clear_cache(&self) -> crate::Result<()> {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage.clear_cache(),
-            GridstoreVariant::Serverless(storage) => storage.clear_cache(),
+            GridstoreVariant::AppendOnly(storage) => storage.clear_cache(),
         }
     }
 }
@@ -308,7 +308,7 @@ impl<V, S: UniversalWrite + 'static> Gridstore<V, S> {
     fn as_dynamic(&self) -> &DynamicGridstore<V, S> {
         match &self.variant {
             GridstoreVariant::Dynamic(storage) => storage,
-            GridstoreVariant::Serverless(_) => panic!("storage is not in dynamic mode"),
+            GridstoreVariant::AppendOnly(_) => panic!("storage is not in dynamic mode"),
         }
     }
 
@@ -316,15 +316,15 @@ impl<V, S: UniversalWrite + 'static> Gridstore<V, S> {
     fn as_dynamic_mut(&mut self) -> &mut DynamicGridstore<V, S> {
         match &mut self.variant {
             GridstoreVariant::Dynamic(storage) => storage,
-            GridstoreVariant::Serverless(_) => panic!("storage is not in dynamic mode"),
+            GridstoreVariant::AppendOnly(_) => panic!("storage is not in dynamic mode"),
         }
     }
 
-    /// Get the inner serverless storage, panics if the storage is in another mode.
-    fn as_serverless(&self) -> &ServerlessGridstore<V, S> {
+    /// Get the inner append-only storage, panics if the storage is in another mode.
+    fn as_append_only(&self) -> &AppendOnlyGridstore<V, S> {
         match &self.variant {
-            GridstoreVariant::Dynamic(_) => panic!("storage is not in serverless mode"),
-            GridstoreVariant::Serverless(storage) => storage,
+            GridstoreVariant::Dynamic(_) => panic!("storage is not in append-only mode"),
+            GridstoreVariant::AppendOnly(storage) => storage,
         }
     }
 }
